requirements.txt
# --- Framework Web ---
fastapi==0.115.0           # Framework backend rápido e assíncrono
uvicorn==0.30.6            # Servidor ASGI para rodar a aplicação FastAPI

# --- Criptografia ---
cryptography==43.0.0       # Geração de chaves RSA e funções criptográficas
pycryptodome==3.20.0       # Implementação do algoritmo IDEA e outras cifras

# --- Autenticação e Segurança ---
passlib==1.7.4             # Hash seguro de senhas (SHA-256 + salt)
python-jose==3.3.0         # JWT tokens (opcional para autenticação baseada em token)
bcrypt==4.0.1
pyjwt>=2.8.0              # Algoritmo bcrypt (opcional para senhas mais seguras)

# --- Banco de Dados ---
sqlalchemy==2.0.34         # ORM para manipulação de dados
databases==0.9.0           # Conexão assíncrona com banco de dados
psycopg2-binary==2.9.9     # Driver PostgreSQL (caso opte por Postgres)
sqlite-utils==3.35         # Utilitário útil se optar por SQLite

# --- Utilidades ---
python-dotenv==1.0.1       # Carregar variáveis de ambiente de um arquivo .env

# --- Testes ---
pytest==8.3.2              # Testes automatizados
pytest-asyncio==0.24.0     # Suporte a testes assíncronos
coverage==7.6.1            # Medição de cobertura de testes

# --- Tipagem e Qualidade ---
mypy==1.11.2               # Verificação de tipos
ruff==0.6.8                # Linter rápido e eficiente
